\name{LeviRmisc-package}
\alias{LeviRmisc-package}
\alias{LeviRmisc}
\docType{package}
\title{Levi's miscellaneous handy R functions}
\description{Just a collection of small but handy functions.}
\details{
\tabular{ll}{Package: \tab LeviRmisc\cr
Maintainer: \tab Levi Waldron <levi.waldron@hunter.cuny.edu>\cr
Author: \tab Levi Waldron and Markus Riester\cr
Version: \tab 0.16.4\cr
License: \tab GPL (>=2.0)\cr
Title: \tab Levi's miscellaneous handy R functions\cr
Imports: \tab multtest, cvTools, Biobase, genomes, XML, gtools, GEOquery\cr
Suggests: \tab jetset, annotate, affy, gplots, GEOmetadb\cr}
}
\author{Levi Waldron and Markus Riester}

\keyword{ package }


